feat: make integrator use modular
Instead of a fixed solver for all ODEs, now the numerical
integrator can be different depending on the problem at hand.
Each problem in the problem set was given one of three possible
integrators: OwrenZen3 for the simpler models; Vern9 for more
complex but non-stiff models; Rodas5 for stiff models.

Signed-off-by: Andre Mello <andre.tgmello@gmail.com>